Update demo code
